% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aktualizuj.baze.R
\name{aktualizuj.baze}
\alias{aktualizuj.baze}
\title{Aktualizacja całej bazy danych}
\usage{
aktualizuj.baze(page_name, how_many = 100, baza, il.watkow = 20,
  il.prob = 60, przerwa = 2)
}
\arguments{
\item{page_name}{Nazwa strony z facebooka, z której będą pobierane posty.}

\item{how_many}{Liczba postów do pobrania (domyślnie 100).}

\item{baza}{Ścieżka do bazy danych lub katalogu gdzie ma zostać utworzona baza.}

\item{il.watkow}{Liczba wątków użytych do przetwarzania danych.}

\item{il.prob}{Liczba prób ściągnięcia danych z facebooka, poprawienia ortografii w
postach lub dokonania rozkładu. Zdarza się, że nie uda się tego zrobić za pierwszym 
razem. Domyślna wartość to 59 powtórzeń.}

\item{przerwa}{Liczba sekund przerwy między kolejnymi próbami aktualizacji bazy.}
}
\description{
Aktualizuje całą bazę danych o posty z facebooka od daty, która jako 
ostatnia widnieje w bazie dla danego banku. Funkcja działa na wielu wątkach.
}

