% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aktualizuj.dane.R
\name{aktualizuj.dane}
\alias{aktualizuj.dane}
\title{Aktualizacja tabeli `dane`}
\usage{
aktualizuj.dane(baza, dane)
}
\arguments{
\item{baza}{Ścieżka do bazy danych lub katalogu gdzie ma zostać utworzona baza.}

\item{dane}{Ramka danych w formie zgodnej ze zwracaną przez funkcję read.page().}
}
\description{
Aktualizuje tabelę `dane` w podanej bazie danych.
}

