% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aktualizuj.rozklady.R
\name{aktualizuj.rozklady}
\alias{aktualizuj.rozklady}
\title{Aktualizacja tabeli `rozklady`}
\usage{
aktualizuj.rozklady(baza, ramka)
}
\arguments{
\item{baza}{Ścieżka do bazy danych lub katalogu gdzie ma zostać utworzona baza.}

\item{ramka}{Ramka danych w formie zwracanej przez funkcję ramka().}
}
\description{
Aktualizuje tabelę `rozklady` w podanej bazie danych.
}

