% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfy_tydzien.R
\docType{data}
\name{cdfy_tydzien}
\alias{cdfy_tydzien}
\title{Dystrybuanty empiryczne dla dni roboczych.}
\format{list}
\usage{
cdfy_tydzien
}
\description{
Obiekt zawiera dystrybuanty empiryczne rozkładu występowania słów kluczowych dla awarii podczas dni roboczych (poniedziałek - piątek).
Słowa kluczowe to "problem", "działać", "utrudnienie", "reklamacja", "awaria", "błąd".
}

