% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfy_weekend.R
\docType{data}
\name{cdfy_weekend}
\alias{cdfy_weekend}
\title{Dystrybuanty empiryczne dla weekendu.}
\format{list}
\usage{
cdfy_weekend
}
\description{
Obiekt zawiera dystrybuanty empiryczne rozkładu występowania słów kluczowych dla awarii podczas weekendu (sobota - niedziela).
Słowa kluczowe to "problem", "działać", "utrudnienie", "reklamacja", "awaria", "błąd".
}

