% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.post.R
\name{get.post}
\alias{get.post}
\title{Pobieranie danych z Facebooka}
\usage{
get.post(posts_id, token, il.watkow = 1)
}
\arguments{
\item{posts_id}{Wektor zawierający numery id postów do pobrania.}

\item{token}{Token, umożliwiający pobieranie danych z Facebooka.}

\item{il.watkow}{Liczba watków, na których będzie uruchamiona funkcja.}
}
\value{
Lista o długości równej długości posts_id. Każdy element jest listą
o 3 elementach:
\itemize{
\item informacje o poście (autor, data utworzenia, id, liczba lików,
 liczba komentarzy, liczba udostępnień),
\item ramka danych z nazwami i id osób, które polubiły post,
\item ramka danych dotycząca komentarzy (autor, wiadomość, czas utworzenia, id).
}
}
\description{
Pobiera dane dotyczące postów o numerach posts_id. Dla każdego posta pobierane
są informacje o poście, polubieniach i komentarzach. Funkcja używa wielu wątków.
}

