% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.comments.R
\name{merge.comments}
\alias{merge.comments}
\title{Lączy post i komentarze w jedną ramkę}
\usage{
\method{merge}{comments}(element, page_name)
}
\arguments{
\item{element}{Lista nazwana zawierająca:
\itemize{
\item post - ramka danych zawierająca informacje o poście (autor, data utworzenia, id, liczba lików,
 liczba komentarzy, liczba udostępnień),
\item likes - ramka danych z nazwami i id osób, które polubiły post,
\item comments - ramka danych dotycząca komentarzy (autor, wiadomość, czas utworzenia, id).
}}

\item{page_name}{Nazwa strony}
}
\value{
Ramka danych
}
\description{
Służy do przetworzenia wyniku działania funkcji getPost z pakietu Rfacebook
do ramki danych.
}

