% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policz.since.R
\name{policz.since}
\alias{policz.since}
\title{dla danej strony sprawdza w bazie date ostatniego postu}
\usage{
policz.since(page_name, baza)
}
\description{
page_name - nazwa strony
baza - sciezka do bazy lub gdzie ma powstac nowa
}

