% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popraw.ortografie.R
\name{popraw.ortografie}
\alias{popraw.ortografie}
\title{Poprawia ortografie, il.prob i przerwa to zmienne ktore definiuja ilosc prob
polaczenia z serwerem i dlugosc przerwy pomiedzy kolejnymi probami (w sek)}
\usage{
popraw.ortografie(dane, il.watkow, il.prob, przerwa)
}
\description{
dane - pobrane z facebooka funkcja read.page(), ramka
il.watkow - argument funkcji makeCluster()
}
\details{
zwraca wektor napisow
}

