% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramka.R
\name{ramka}
\alias{ramka}
\title{rozklad - policzony przez serwer, jak na zajeciach, funkcja policz.rozklad()
dane - pobrane z facebooka funkcja read.page(), ramka}
\usage{
ramka(rozklad, dane)
}
\description{
zwraca ramke danych
}

