% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.page.R
\name{read.page}
\alias{read.page}
\title{scraping z Facebooka, wczytuje posty z komentarzami ze strony page_name
       i zwraca ramke danych
page_name - nazwa strony
how_many - ile postow ze strony czytac
token - token lub plik z pakietu (plik fb_oauth), ktory trzeba wczytac load("fb_oauth")}
\usage{
read.page(page_name, how_many, token, since, il.watkow)
}
\description{
scraping z Facebooka, wczytuje posty z komentarzami ze strony page_name
       i zwraca ramke danych
page_name - nazwa strony
how_many - ile postow ze strony czytac
token - token lub plik z pakietu (plik fb_oauth), ktory trzeba wczytac load("fb_oauth")
}

