% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policz.rozklad.R
\name{policz.rozklad}
\alias{policz.rozklad}
\title{Poprawia ortografie, proby i przerwa to zmienne ktore definiuja ilosc prob
polaczenia z serwerem i dlugosc przerwy pomiedzy kolejnymi probami (w sek)}
\usage{
policz.rozklad(dane, il.watkow, il.prob, przerwa)
}
\arguments{
\item{dane}{Ramka danych, zawierająca kolumnę napisów body}

\item{il.watkow}{Liczba wątków, na których będzie wykonywana funckja.}

\item{il.prob}{Liczba prób w przypadku niepowodzenia.}

\item{przerwa}{Przerwa pomiedzy probami.}
}
\value{
Zwraca rozklad zdań w postach w postaci listy list.
}
\description{
Rozkłady zdań w treści postów.
}
\details{
Dla ramki danych zaweirającej kolumnę body listę rozkładów zdań w tej
ramce. Korzysta z serwsisu:
https://ec2-54-194-28-130.eu-west-1.compute.amazonaws.com/ams-ws-nlp/rest/nlp/single
W przypadku niepowodzenia próbuje wiele razy, każda próba jest odzielona przerwą.
}

