% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policz.since.R
\name{policz.since}
\alias{policz.since}
\title{Sprawdzanie daty ostatniego postu}
\usage{
policz.since(page_name, baza)
}
\arguments{
\item{page_name}{Nazwa strony.}

\item{baza}{Ścieżka do bazy.}
}
\value{
Zwracany jest wektor jednoelementowy zawierający datę, jeśli taka istnieje.
}
\description{
Funkcja ma za zadanie dla danej strony sprawdzić w bazie date ostatniego postu.
}

