% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popraw.ortografie.R
\name{popraw.ortografie}
\alias{popraw.ortografie}
\title{Poprawianie ortografii}
\usage{
popraw.ortografie(dane, il.watkow, il.prob, przerwa)
}
\arguments{
\item{dane}{Zbiór danych dla którego chcemy poprawić ortografię.}

\item{il.watkow}{Liczba wątków, na których będzie uruchomiona funkcja.}

\item{il.prob}{Liczba prób połączenia z serwerem.}

\item{przerwa}{Czas w [s] który po nieudanej próbie połączenia z serwerem odczekamy zanim nawiążemy kolejne połączenia.}
}
\value{
Zwracany jest wektor napisów zapisany w kodowaniu utf-8. Jeśli próba połączenia się
z serwerem nie powiedzie się zwracany jest stosowny napis.
}
\description{
Poprawiana jest ortografia w danych. Funkcja używa wielu wątków.
}

