% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramka.R
\name{ramka}
\alias{ramka}
\title{Tworzenie ramki danych}
\usage{
ramka(rozklad, dane)
}
\arguments{
\item{rozklad}{Rozkład zgodnie z którym ma być stworzona ramka danych. Podajemy jako listę.}

\item{dane}{Zbiór danych dla którego chcemy stworzyć razkład. Podajemy jako ramkę danych.}
}
\value{
Zwracana jest ramka danych składająca się z napisów zapisanych w kodowaniu utf-8.
}
\description{
Funkcja tworzy ramkę danych stworzoną według podanego rozkładu. Jako rozkład można wykorzystać
inną pomocną funkcję z pakietu "policz.rozklad()".
}

