% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.page.R
\name{read.page}
\alias{read.page}
\title{Wczytywanie danych z Facebooka}
\usage{
read.page(page_name, how_many, token, since, il.watkow)
}
\arguments{
\item{page_name}{Nazwa strony.}

\item{how_many}{Ile postów ze strony mamy pobrać.}

\item{token}{Token lub plik z pakietu (plik fb_oauth), który trzeba wczytać.}

\item{since}{Data od której chcemy pobierać informacje.}

\item{il.watkow}{Na ilu wątkach mają się pobierać dane.}
}
\value{
Zwracana jest ramka danych składająca się z napisów.
}
\description{
Funkcja read.page() wczytuje posty wraz z komentarzami ze strony page_name
i zwraca ramkę danych.
}

